% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sui_js}
\alias{sui_js}
\title{Inject javascript needed by the sui_shinymod modules}
\usage{
sui_js()
}
\value{
A script tag
}
\description{
Inject javascript needed by the sui_shinymod modules
}
\seealso{
\code{\link{sui_shinymod_ui}}
}
