% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{sui_shinymod_ui}
\alias{sui_shinymod_ui}
\alias{sui_shinymod_server}
\title{A Shiny module for internationalization}
\usage{
sui_shinymod_ui(id)

sui_shinymod_server(input, output, session, csv_path = NULL)
}
\arguments{
\item{id}{string: the id of the shiny element}

\item{input}{: see \code{\link[shiny]{callModule}}}

\item{output}{: see \code{\link[shiny]{callModule}}}

\item{session}{: see \code{\link[shiny]{callModule}}}

\item{csv_path}{string: path to the languages CSV file. See \code{\link{sui_translator}} for details}
}
\description{
A Shiny module for internationalization
}
\examples{
\dontrun{
  ## in ui function
  sui_js()
  sui_shinymod_ui(id = "lang")

  ## in server function
  callModule(sui_shinymod_server, id = "lang", csv_path = "optional/path/to/your/file")
}
}
