% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{sui_translator}
\alias{sui_translator}
\title{Construct i18n translator}
\usage{
sui_translator(
  to,
  csv_path = system.file("extdata/su_translations.csv", package = "sui18n")
)
}
\arguments{
\item{to}{string: language to translate to, defaults to "en" (i.e. no translation)}

\item{csv_path}{string: path to the translation CSV file. \code{csv_path} can also be provided directly as a data.frame}
}
\value{
A list object. The translation functionality is provided by the \code{t} component. The \code{t} function will first look for an exact match between the input text and the reference data. If no exact match is found, further matching behaviour is controlled by the following options, which are applied in order:
\itemize{
 \item \code{underscores_as_spaces} (default = \code{TRUE}) - underscores in the input will be converted to spaces
 \item \code{as_regexp} (default = \code{FALSE}) - the input is treated as a regular expression. Note that the regular expression must match the entire input string
 \item \code{ignore_case} (default = \code{TRUE}) - allow case-insensitive matching
 \item \code{allow_punct} (default = \code{TRUE}) - allow additional punctuation and space characters at the beginning and end of the input
}
The \code{t} function also takes an \code{output_case} string parameter, which can take values "match" (attempt to match the case style of the output to that of the input), "upper" (output upper case), "firstupper" (first character opf the output to upper case), "title" (title case), "lower", (lower case), or "as-is" (output will be whatever has been entered into the translations file). By default this is "match" except when translating to German ("as-is"), but this behaviour might change.
}
\description{
The CSV data should contain one column per language.
}
\examples{

tr <- sui_translator()
tr$t("hello")
tr$set_target("fr")
tr$t("hello")
tr$t("Hello")
tr$t("Hello!")

}
