% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{sui_translator}
\alias{sui_translator}
\title{Construct SU i18n translator}
\usage{
sui_translator(
  to,
  csv_path = system.file("extdata/su_translations.csv", package = "sui18n")
)
}
\arguments{
\item{to}{string: language to translate to, defaults to "en" (i.e. no translation)}

\item{csv_path}{string: path to the translation CSV file}
}
\value{
A list
}
\description{
Construct SU i18n translator
}
\examples{

tr <- sui_translator()
tr$t("hello")
tr$set_target("fr")
tr$t("hello")
tr$t("Hello")
tr$t("Hello!")

}
